requirements.txt
python-docx
docxtpl
Flask
flask-cors
